gres underflow error fix
this bug exists with all select plugins. if a job has been allocated
gres and the gres have either topology or type information and the
slurmctld daemon restarts (while the job is running), then when the
job ends gres underflow errors will be generated. the problem is
due to the slurmctld not having gres topology or type information
available at restart time so that it can not update counters. the
overhead of updating those counters at node registration time is
high, so we just avoid generating the errors in this case.
note: this bug is not specific to cons_tres and  exists in
earlier versions of slurm.